Update prepocess.py
Made several changes. Modified preprocess() such that the data is read into a pandas DataFrame, split, the unique columns for one-hot encoding are found (in a parallel way), then the one-hot encoding occurs after the new columns are added to the DataFrame. Added a new internal function _get_new_column() which is used for the aforementioned step necessary for one-hot encoding. Lastly, a merge_dataframe function was written to merge DataFrames.